🎨 :harden backend — int64 money, error taxonomy, refresh token, REST path params
Security & middleware:
- remove CSRF middleware (stateless Bearer API) and add CSP header
- fix role check constraint to (0,1) with DEFAULT 1 (staff registration)

Money handling:
- migrate all money fields float64 -> int64 rupiah (domain/repo/usecase/DTO/pdf)
- include Midtrans PPN 11% in passthrough fee (VA 4440, QRIS 0.777%)
- snapshot purchase_price & product_name on transaction detail

Error handling & consistency:
- add ErrNotFound/ErrInvalidID/Duplicate taxonomy + writeError helper
- map duplicate invoice/order_id -> 409, not-found -> 404
- normalize repo not-found to domain.NotFound

Auth & API:
- add POST /auth/refresh with atomic token rotation (GETDEL)
- convert DELETE/PUT to /:id path params (breaking)

Fixes:
- report timezone Asia/Jakarta, bigint casts, enum.Hutang refs
- GetTransaction total profit & TransactionID; fee-on-fee; debt upsert filter
- JWT TTL unit conversion (seconds) in composition root
- upgrade fiber v3.5.0 + gorm v1.31.2

BREAKING CHANGE: DELETE/PUT now use /:id path param instead of body ID

+++ internal/usecase/product_usecase_test.go
@@ -118,7 +118,7 @@ func TestUpdateProductShopWithLock_BuildsPartialFieldMap(t *testing.T) {
 
 	id := uuid.New()
 	name := "Gula Baru"
-	price := 15000.0
+	price := int64(15000)
 	req := &requestdto.UpdateProduct{ID: id.String(), ProductName: &name, SellingPrice: &price}
 
 	if err := u.UpdateProductShopWithLock(context.Background(), req, 5); err != nil {
